backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0

# LangChain stack
langchain==0.3.23
langchain-core>=0.3.51,<1.0.0
langchain-pinecone==0.2.5
langchain-ollama==0.3.1

# Pinecone
pinecone-client==2.2.4

# Ollama
httpx>=0.25.2

# Alpaca
alpaca-py==0.39.2

# General
python-dotenv==1.0.0
pydantic
aiohttp>=3.8.3,<4.0.0
pypdf==3.17.1
pypdf2==3.0.1
pandas>=1.3.5,<2.0.0
numpy==1.26.4
python-multipart==0.0.6
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0 
